requirements.txt
ollama
python-docx 